Send RPM data to the WebSocket!

+++ src/hall-effect.h
@@ -4,12 +4,12 @@ class HallEffect {
    private:
     volatile int count;
     unsigned long lastReading;
-    void (*callback)(float);
+    void (*callback)(double);
 
     void interrupt();
 
    public:
     void setup();
     void loop();
-    void onData(void (*callback)(float));
+    void onData(void (*callback)(double));
 };
